Rename abi files so they dont have a json ext
If they have a json or yaml extension, then sls imports them as objects instead of strings, and the schema validation for abi to be a string fails.

A good solution would be to have "abi" be a string or object, but in sls, ajv runs in strict mode and disallows union types.

A workaround would be to have a property "abi" of type string and an "abiObject" of type object, and manually check that either (but not both) is present. Oh well.